Move app, templates and static files into src directory
